Don't autoload unrelated proxy classes.
