+++ vnet/lib/vnet/api_rpc.rb
@@ -8,7 +8,11 @@ module Vnet
     raise "Could not load 'vnet/api_rpc', api proxy already loaded."
   end
 
-  use_api_proxy = :direct
+  use_api_proxy = :rpc
+
+  module NodeApi
+    autoload :RpcProxy, 'vnet/node_api/rpc_proxy'
+  end
 
   Vnet::NodeApi.set_api_proxy(Vnet::NodeApi::RpcProxy.new)
