+++ vnet/lib/vnet/node_api.rb
@@ -14,9 +14,53 @@ module Vnet
         @proxy = new_proxy
       end
     end
+
+    class Proxy
+      def method_missing(class_name, *args, &block)
+        # The const_defined search constructs all the NodeApi::*
+        # classes, even if we are using 'rpc', and as such can't do a
+        # search to validate the class names.
+        if class_name.present? && args.empty?
+          _call_class.new(class_name).tap do |call|
+            define_singleton_method(class_name){ call }
+          end
+        else
+          super
+        end
+      end
+
+      protected
+
+      def _call_class
+        raise NotImplementedError
+      end
+
+      class Call
+        def initialize(class_name)
+          @class_name = class_name
+        end
+
+        def method_missing(method_name, *args, &block)
+          _call(method_name, *args, &block)
+        rescue => exception
+          raise exception if Vnet::NodeApi.raise_on_error
+          logger.debug("#{exception.class}: #{exception.to_s}\n\t")
+          nil
+        end
+
+        private
+
+        def _call(method_name, *args, &block)
+          raise NotImplementedError
+        end
+
+        def logger
+          Vnet.logger
+        end
+      end
+    end
+
   end
 end
-
-require_relative("node_api/proxies")
 
 Vnet::NodeApi.raise_on_error = true
