+++ vnet/lib/vnet/node_api/rpc_proxy.rb
@@ -1,0 +1,23 @@
+# -*- coding: utf-8 -*-
+
+module Vnet
+  module NodeApi
+    class RpcProxy < Proxy
+      protected
+      class RpcCall < Call
+        def initialize(class_name)
+          super
+          @actor = DCell::Global[:rpc] or raise "rpc not found in DCell::Global"
+        end
+
+        def _call(method_name, *args, &block)
+          @actor.execute(@class_name, method_name, *args, &block)
+        end
+      end
+
+      def _call_class
+        RpcCall
+      end
+    end
+  end
+end
